Separated embed functions into separate file
